Tries to create a value from fields parents if the field is not registered.
When project defines a lot of custom fields, it can be a hassle to
register all of them manually.
Quite often the registered sample data for the field the custom field
was inherited from will work just as well. Hence, it makes sense to try
them first before reporting an error.